back-end/requirements.txt
albumentations==1.1.0
coremltools==5.2.0
Django==3.2.13
ipython==7.32.0
matplotlib==3.5.1
onnxruntime==1.11.0
opencv_python==4.5.5.64
openvino==2022.1.0
pafy==0.5.5
pandas==1.3.5
Pillow==9.1.0
protobuf==3.20.1
psutil==5.9.0
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
tflite_runtime==2.8.0
thop==0.0.31-2005241907
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
val==0.6
